requirements.txt
# ── Geply Dependencies ──

# API Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database — SQLite (zero setup)
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0

# Validation & Settings
pydantic==2.10.4
pydantic-settings==2.7.1

# Auth — PyJWT (actively maintained, replaces abandoned python-jose)
PyJWT[crypto]==2.10.1
bcrypt==4.2.1

# File Processing
python-docx==1.1.2
pypdf==5.1.0
aiofiles==24.1.0

# HTTP Client
httpx==0.28.1

# Utilities
python-dateutil==2.9.0
structlog==24.4.0

# TTS
edge-tts==6.1.12

# Deepgram Voice Agent (WebSocket client)
websockets>=12.0
